Significant changes relating to allocation and deallocation of derived type objects. The changes only really affect derived type instances allocated from fortran. Changes were made to how the reference counting is done. Now, the fortran pointer itself claims a reference, which is released during the call to Releasetypename. The fscalar->data elements also claim references. An object is only deleted when its reference count goes to zero.
When a derived type allocated in fortran is accessed from python, the
appropriate fscalar->data grabs a reference. This reference is only
given up when fscalar->data is reassigned - this only happens if the
pointer is again referenced from python (to get the new object). This
means that if an object is accessed from python, the object will only
actually be deleted when the pointer is accessed again, even when the
appropriate release is called.  This could potentially cause problems if
the undeleted objects are large.

Note that the garbage collection code is not fully tested, and there does
appear to be a small memory leak remaining.