In setup.py, for Darwin, fixed data install location
When setting up the data scheme for data files, on Darwin, the
location must be explicitly set. In Darwin, after the macros are
expanded in the scheme install locations, the purelib and platlib
are modified to go into the /Library root, but data is not. The
data files were being placed into the /System/Library root. The
data is now explicitly set to go into /Library.