Fixed FARRAY_FROMOBJECT
There was a change in numpy when creating arrays. Multidimensional
arrays that have only one dimension with a length greater than one now
have the F_CONTIGUOUS flag set to true. This was confusing FARRAY_FROMOBJECT.
That routine was rewritten to use the standard numpy PyArray_CheckFromAny, with the NPY_ARRAY_F_CONTIGUOUS set. This gets the strides and dimensions in the
order that Forthon needs.